Remove http.HttpReads backwards-compatible type
It didn't work very elegantly, and also could lead to non-backwards compatibility
being hidden. This way, we force source recompilation and therefore implicit
resolution
