+++ src/main/scala/uk/gov/hmrc/play/http/HttpDelete.scala
@@ -19,6 +19,7 @@ package uk.gov.hmrc.play.http
 import play.api.http.HttpVerbs.{DELETE => DELETE_VERB}
 import uk.gov.hmrc.play.audit.http.{HeaderCarrier, HttpAuditing}
 import uk.gov.hmrc.play.http.logging.{MdcLoggingExecutionContext, ConnectionTracing}
+import uk.gov.hmrc.play.http.reads.HttpReads
 import scala.concurrent.Future
 import MdcLoggingExecutionContext._
