+++ src/main/scala/uk/gov/hmrc/play/http/HttpPost.scala
@@ -21,6 +21,7 @@ import play.api.http.HttpVerbs.{POST => POST_VERB}
 import uk.gov.hmrc.play.audit.http.{HeaderCarrier, HttpAuditing}
 import uk.gov.hmrc.play.http.logging.{MdcLoggingExecutionContext, ConnectionTracing}
 import MdcLoggingExecutionContext._
+import uk.gov.hmrc.play.http.reads.HttpReads
 
 import scala.concurrent.Future
