+++ src/main/scala/uk/gov/hmrc/play/http/reads/HtmlHttpReads.scala
@@ -17,6 +17,7 @@
 package uk.gov.hmrc.play.http.reads
 
 import play.twirl.api.Html
+import uk.gov.hmrc.play.http.reads.HttpReads
 
 trait HtmlHttpReads {
   def bodyToHtml = HttpReads[Html] { (method, url, response) => Html(response.body) }
